# openapi.yaml
openapi: 3.0.0
info:
  title: REST API для ресторана, Laravel, PHP
  description: iz4GsmrfBVPARkSz9ggvm2QuuJtzQGUrkXQSojsm     admin@admin.ru admin    1111
  version: '0.1'
servers:
  - url: http://lar-test-sw.ru/api/
    description: Dev server
paths:
...
  /dishes/{id}/update:
    put:
      summary: Обновление данных блюда
      description: Доступно только суперадминам и админам
      tags:
        - Dishes
      operationId: updateDish
      parameters:
        - in: path
          name: id
          example: 15
          description: ID блюда
          required: true
          schema:
            type: integer
      requestBody:
        description: Могут быть любые из предложенных полей. Если полей нет, запрос тоже пройдет успешно, ничего не  изменив
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreDishRequest'
      responses:
        '200':
          description: Сообщение об успешном обновлении
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OKRequest"
        '404':
          description: Ошибка поиска
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error404"
        '422':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorValidation"
...
components:
  schemas:
...
    ErrorValidation:
      type: object
      properties:
        message:
          type: string
          example: Описание ошибки
        errors:
          type: object
          example:
            field1:
              - Необходимо поле названия.
            field2:
              - Айди не существует.
            field_N:
              - Необходимо ввести пароль ещё раз.
          description: Список ошибок и список полей, в которых они есть
...
    StoreDishRequest:
      type: object
      properties:
        name:
          type: string
          example: Биг-мак
        img:
          type: string
          example: Image File
          description: Картинка блюда
        composition:
          type: string
          example: Две мясные котлеты гриль, специальный соус, сыр, огурцы, салат и лук, всё на булочке с кунжутом
          description: Картинка категории
        calories:
          type: integer
          example: 500
        price:
          type: integer
          example: 200
        category_id:
          type: integer
          example: 3
...
    Error404:
      type: object
      properties:
        message:
          type: string
          example: "[Объект] не найден(o/a)"
    OKRequest:
      type: object
      properties:
        message:
          type: string
          example: "Успешно"